requirements.txt
luma==2.1.5
luma.led_matrix==1.6.1
spotipy~=2.19.0
python-dotenv~=0.20.0